Replace background script with event listener
The new approach is simpler and no longer fires the content script
multiple times - the page loads itself dynamically and while the
background script listened to XMLHttpRequests it was not able to
distinguish between them.
Make the values for minimum profit and discount percentage constants
on the common functions script.