Restructure repo, add makefile, keep both browsers manifests
Add makefile

The only obstacle to have cross browser support is the manifest
files, firefox requires v2 and chrome v3, thus I added two versions
of the manifests and keep the remaining code separate.

To load the extension during development, its as simple as moving
one of the manifests into the folder with the remaining code.